Fix starting message (#20)
the starting message was supposed to be cleared on reception of a VS
Code specific notifications. This specific notification system was
implemented in jdt.ls. I removed all specific protocols from Apache
Camel as we don't need them
for now.
I changed to update the message to "started" when the "connection" has
been successful with the Language Server.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>