FUSETOOLS2-1591 - upgrade minimal version to 1.61.0
it was broken with previous version. This an already relatively old
version which should be supported by Theia/Che/CRW.

It required to adapt the UI test for unclear reasons the Dialog.close is
not working with 1.66.0 when called with specific version on CI, neither
the File -> Revert File fo rsome menus.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>