FUSETOOLS2-2103 - Create a Camel Quarkus Project
- validation of gav partial

It is picking all Camel files in the root workspace and copying them in
the src/main/resources:camel folder. if there are none, the structure is
still created and it is up to the users to create it.

I think will keep for another iteration:
- allow to open in another workspace the created project
- a more complete validation

/!\ When a *.camel.xml is used, the generated project will fail when
launching it because Camel Quarkus is not using a recent enough version
of Camel which contains the fix for the extension name.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>