FUSETOOLS2-2540 - avoid flakiness in tests due to file cannot be saved
it was resulting to next test to not be able to activate the extension.
this test doesn't need to save the file when closing the editor as the
file is deleted right after. Note that I have not found why when trying
to save there was an error stating that the file does not exist.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>